Move Instances objects to arrow-core
