+++ modules/core/arrow-core/src/main/kotlin/arrow/core/boolean.kt
@@ -1,0 +1,3 @@
+package arrow.core
+
+object BooleanInstances
