+++ modules/core/arrow-core/src/main/kotlin/arrow/core/map.kt
@@ -1,0 +1,5 @@
+package arrow.core
+
+object MapInstances
+
+object SortedMapInstances
