+++ modules/core/arrow-core/src/main/kotlin/arrow/core/set.kt
@@ -1,0 +1,5 @@
+package arrow.core
+
+object SetInstances
+
+object SortedSetInstances
