+++ modules/core/arrow-instances-core/src/main/kotlin/arrow/instances/boolean.kt
@@ -1,5 +1,6 @@
 package arrow.instances
 
+import arrow.core.BooleanInstances
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Show
 
@@ -19,11 +20,8 @@ interface BooleanEqInstance : Eq<Boolean> {
 //fun Boolean.Companion.eq(): Eq<Boolean> =
 //  object : BooleanEqInstance {}
 
-object BooleanInstances {
+fun BooleanInstances.show(): Show<Boolean> =
+  object : BooleanShowInstance {}
 
-  fun show(): Show<Boolean> =
-    object : BooleanShowInstance {}
-
-  fun eq(): Eq<Boolean> =
-    object : BooleanEqInstance {}
-}
+fun BooleanInstances.eq(): Eq<Boolean> =
+  object : BooleanEqInstance {}
