--- modules/core/arrow-instances-data/src/main/kotlin/arrow/instances/list.kt
@@ -1,3 +1,0 @@
-package arrow.instances
-
-object ListInstances
